+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+Breaking
+  * AZ-302: Replace diagnostics agent with Azure Monitor agent
+
 Changed
   * AZ-572: Revamp examples and improve CI